Refactoring
Move some functions to shared crate modules, they are required by other
commands.
